gif_creator/requirements.txt
streamlit
pillow
streamlit-login-auth-ui
moviepy